# src/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Universal Resolver Driver LTO
  description: Universal Resolver Driver LTO
  contact:
    name: Sphereon B.V.
    email: dev@sphereon.com
  version: 1.0.0
servers:
 - url : /
paths:
  /1.0/identifiers/{identifier}:
    get:
      summary: Resolve a DID or other identifier.
      operationId: resolve
      parameters:
      - name: identifier
        in: path
        description: A DID or other identifier to be resolved.
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: did:lto:testnet:3N51gbw5W3xvSkcAXtLnXc3SQh2m9e6TBcy
#      - name: Accept
#        in: header
#        description: The requested MIME type of the DID document or DID resolution
#          result.
#        required: false
#        style: simple
#        explode: false
#        schema:
#          type: string
#        examples:
#          application/did+json:
#            description: MIME type of a DID document (JSON representation)
#            value: application/json
#          application/did+ld+json:
#            description: MIME type of a DID document (JSON-LD representation)
#            value: application/did+ld+json
#          application/ld+json;profile=\"https://w3c-ccg.github.io/did-resolution/\":
#            description: MIME type of a DID resolution result
#            value: application/ld+json;profile=\"https://w3c-ccg.github.io/did-resolution/\"
      responses:
        "200":
          description: successfully resolved!
          content:
            application/did+json:
              schema:
                $ref: '#/components/schemas/ResolutionResult'

            application/did+ld+json:
              schema:
#                type: object
#                description: The DID document (JSON-LD representation).
                $ref: '#/components/schemas/ResolutionResult'

            application/ld+json;profile=\"https://w3c-ccg.github.io/did-resolution/\":
              schema:
                $ref: '#/components/schemas/ResolutionResult'
        "400":
          description: invalid input!
        "500":
          description: error!
      x-swagger-router-controller: didController
components:
  schemas:
    ResolutionResult:
      type: object
      properties:
        didDocument:
          type: object
          example:
            '@context': https://www.w3.org/2019/did/v1
            id: did:lto:3JuijVBB7NCwCz2Ae5HhCDsqCXzeBLRTyeL
            publicKey:
            - id: did:lto:3JuijVBB7NCwCz2Ae5HhCDsqCXzeBLRTyeL#key-1
              type: Ed25519VerificationKey2018
              publicKeyBase58: H3C2AVvLMv6gmMNam3uVAjZpfkcJCwDwnZn6z3wXmqPV
        resolverMetadata:
          type: object
        methodMetadata:
          type: object
      description: The DID resolution result.
      example:
        '@context': https://www.w3.org/2019/did/v1
        id: did:lto:testnet:3N51gbw5W3xvSkcAXtLnXc3SQh2m9e6TBcy
        publicKey:
          - id: did:lto:3JuijVBB7NCwCz2Ae5HhCDsqCXzeBLRTyeL#key-1
            type: Ed25519VerificationKey2018
            publicKeyBase58: H3C2AVvLMv6gmMNam3uVAjZpfkcJCwDwnZn6z3wXmqPV